+++ src/Oro/Bundle/EntityBundle/Provider/ChainVirtualRelationProvider.php
@@ -2,16 +2,36 @@
 
 namespace Oro\Bundle\EntityBundle\Provider;
 
-class ChainVirtualRelationProvider extends AbstractChainProvider implements VirtualRelationProviderInterface
+use Oro\Bundle\EntityConfigBundle\Provider\ConfigProvider;
+use Oro\Bundle\EntityExtendBundle\Tools\ExtendHelper;
+
+/**
+ * Delegates building of virtual relations to child providers.
+ */
+class ChainVirtualRelationProvider implements VirtualRelationProviderInterface
 {
+    /** @var iterable|VirtualRelationProviderInterface[] */
+    private $providers;
+
+    /** @var ConfigProvider  */
+    private $configProvider;
+
+    /**
+     * @param iterable|VirtualRelationProviderInterface[] $providers
+     * @param ConfigProvider                              $configProvider
+     */
+    public function __construct(iterable $providers, ConfigProvider $configProvider)
+    {
+        $this->providers = $providers;
+        $this->configProvider = $configProvider;
+    }
+
     /**
      * {@inheritdoc}
      */
     public function isVirtualRelation($className, $fieldName)
     {
-        /** @var VirtualRelationProviderInterface[] $providers */
-        $providers = $this->getProviders();
-        foreach ($providers as $provider) {
+        foreach ($this->providers as $provider) {
             if ($provider->isVirtualRelation($className, $fieldName)) {
                 return true;
             }
@@ -33,18 +53,22 @@ class ChainVirtualRelationProvider extends AbstractChainProvider implements Virt
      */
     public function getVirtualRelations($className)
     {
-        /** @var VirtualRelationProviderInterface[] $providers */
-        $providers = $this->getProviders();
-        $result = [];
-
-        foreach ($providers as $provider) {
-            $virtualRelations = $provider->getVirtualRelations($className);
-            if (!empty($virtualRelations)) {
-                $result = array_merge($result, $virtualRelations);
-            }
+        if (!$this->isEntityAccessible($className)) {
+            return [];
         }
 
-        return $result;
+        $relations = [];
+        foreach ($this->providers as $provider) {
+            $virtualRelations = $provider->getVirtualRelations($className);
+            if (!empty($virtualRelations)) {
+                $relations[] = $virtualRelations;
+            }
+        }
+        if ($relations) {
+            $relations = array_merge(...$relations);
+        }
+
+        return $relations;
     }
 
     /**
@@ -52,11 +76,8 @@ class ChainVirtualRelationProvider extends AbstractChainProvider implements Virt
      */
     public function getTargetJoinAlias($className, $fieldName, $selectFieldName = null)
     {
-        return $this->findProvider($className, $fieldName)->getTargetJoinAlias(
-            $className,
-            $fieldName,
-            $selectFieldName
-        );
+        return $this->findProvider($className, $fieldName)
+            ->getTargetJoinAlias($className, $fieldName, $selectFieldName);
     }
 
     /**
@@ -65,28 +86,30 @@ class ChainVirtualRelationProvider extends AbstractChainProvider implements Virt
      *
      * @return VirtualRelationProviderInterface
      */
-    protected function findProvider($className, $fieldName)
+    private function findProvider(string $className, string $fieldName): VirtualRelationProviderInterface
     {
-        $foundProvider = null;
-        /** @var VirtualRelationProviderInterface[] $providers */
-        $providers = $this->getProviders();
-        foreach ($providers as $provider) {
+        foreach ($this->providers as $provider) {
             if ($provider->isVirtualRelation($className, $fieldName)) {
-                $foundProvider = $provider;
-                break;
+                return $provider;
             }
         }
 
-        if ($foundProvider === null) {
-            throw new \RuntimeException(
-                sprintf(
-                    'A query for relation "%s" in class "%s" was not found.',
-                    $fieldName,
-                    $className
-                )
-            );
-        }
+        throw new \RuntimeException(sprintf(
+            'A query for relation "%s" in class "%s" was not found.',
+            $fieldName,
+            $className
+        ));
+    }
 
-        return $foundProvider;
+    /**
+     * @param string $className
+     *
+     * @return bool
+     */
+    private function isEntityAccessible(string $className): bool
+    {
+        return
+            !$this->configProvider->hasConfig($className)
+            || ExtendHelper::isEntityAccessible($this->configProvider->getConfig($className));
     }
 }