+++ src/Oro/Bundle/EntityConfigBundle/Config/ConfigCacheWarmer.php
@@ -6,9 +6,9 @@ use Doctrine\ORM\Mapping\ClassMetadata;
 use Oro\Bundle\EntityBundle\Provider\VirtualFieldProviderInterface;
 use Oro\Bundle\EntityBundle\Provider\VirtualRelationProviderInterface;
 use Oro\Bundle\EntityConfigBundle\Entity\ConfigModel;
-use Oro\Bundle\EntityExtendBundle\EntityConfig\ExtendScope;
 
 /**
+ * Warms up entity configuration cache.
  * IMPORTANT: A performance of this class is very crucial. Double check a performance during a refactoring.
  */
 class ConfigCacheWarmer
@@ -94,10 +94,8 @@ class ConfigCacheWarmer
             try {
                 $this->loadNonConfigurable();
                 $this->loadVirtualFields();
+            } finally {
                 $this->configModelLockObject->unlock();
-            } catch (\Exception $e) {
-                $this->configModelLockObject->unlock();
-                throw $e;
             }
         }
     }
@@ -246,14 +244,8 @@ class ConfigCacheWarmer
 
     protected function loadVirtualFields()
     {
-        $extendConfigProvider = $this->configManager->getProvider('extend');
-
         $entities = $this->cache->getEntities();
         foreach ($entities as $className => $entityData) {
-            if ($extendConfigProvider->getConfig($className)->is('state', ExtendScope::STATE_NEW)) {
-                continue;
-            }
-
             $virtualFields = $this->virtualFieldProvider->getVirtualFields($className);
             if (!empty($virtualFields)) {
                 foreach ($virtualFields as $fieldName) {